# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Retail POS + Inventory + Barcode System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  cashier: "cashier",
  inventory: "inventory",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  cashier: "cashier",
  inventory: "inventory",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  cashier: "cashier",
  inventory: "inventory",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  password?: string;
  role?: UpdateUserBodyRole;
}

export interface Product {
  id: number;
  name: string;
  title: string;
  price: number;
  category: string;
  stock: number;
  barcode: string;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  title: string;
  price: number;
  category: string;
  stock: number;
}

export interface UpdateProductBody {
  name?: string;
  title?: string;
  price?: number;
  category?: string;
  stock?: number;
}

export interface SaleItem {
  productId: number;
  productName: string;
  barcode: string;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Sale {
  id: number;
  items: SaleItem[];
  totalAmount: number;
  cashierId: number;
  /** @nullable */
  cashierName?: string | null;
  createdAt: string;
}

export type CreateSaleBodyItemsItem = {
  productId: number;
  quantity: number;
};

export interface CreateSaleBody {
  items: CreateSaleBodyItemsItem[];
  cashierId: number;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  lowStockCount: number;
  avgOrderValue: number;
}

export interface SalesChartPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  lowStock?: boolean;
};

export type ListSalesParams = {
  period?: ListSalesPeriod;
  cashierId?: number;
};

export type ListSalesPeriod =
  (typeof ListSalesPeriod)[keyof typeof ListSalesPeriod];

export const ListSalesPeriod = {
  today: "today",
  week: "week",
  month: "month",
  all: "all",
} as const;

export type GetDashboardSummaryParams = {
  period?: GetDashboardSummaryPeriod;
};

export type GetDashboardSummaryPeriod =
  (typeof GetDashboardSummaryPeriod)[keyof typeof GetDashboardSummaryPeriod];

export const GetDashboardSummaryPeriod = {
  today: "today",
  week: "week",
  month: "month",
} as const;

export type GetSalesChartParams = {
  days?: number;
};

export type GetTopProductsParams = {
  period?: GetTopProductsPeriod;
  limit?: number;
};

export type GetTopProductsPeriod =
  (typeof GetTopProductsPeriod)[keyof typeof GetTopProductsPeriod];

export const GetTopProductsPeriod = {
  today: "today",
  week: "week",
  month: "month",
} as const;

export type GetLowStockProductsParams = {
  threshold?: number;
};