# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Retail POS + Inventory + Barcode System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product with auto barcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Get product by barcode (for POS scanning)
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard analytics summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/sales-chart:
    get:
      operationId: getSalesChart
      tags: [dashboard]
      summary: Get sales data for chart (last 7 or 30 days)
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesChartPoint"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /dashboard/low-stock:
    get:
      operationId: getLowStockProducts
      tags: [dashboard]
      summary: Get products with low stock (below threshold)
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Low stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, cashier, inventory]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        price:
          type: number
        category:
          type: string
        stock:
          type: integer
        barcode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - title
        - price
        - category
        - stock
        - barcode
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        price:
          type: number
        category:
          type: string
        stock:
          type: integer
      required:
        - name
        - title
        - price
        - category
        - stock
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        price:
          type: number
        category:
          type: string
        stock:
          type: integer
...
    DashboardSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalProducts:
          type: integer
        lowStockCount:
          type: integer
        avgOrderValue:
          type: number
      required:
        - totalRevenue
        - totalOrders
        - totalProducts
        - lowStockCount
        - avgOrderValue
    SalesChartPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: number
      required:
        - productId
        - productName
        - totalQuantity
        - totalRevenue