# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Retail POS + Inventory + Barcode System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  title: string;
  price: number;
  category: string;
  stock: number;
  barcode: string;
  createdAt: string;
}